pyproject.toml
[project]
name = "auto-ticket-classifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "coverage>=7.13.4",
    "fastapi[standard]>=0.135.1",
    "google-genai>=1.66.0",
    "noisereduce>=3.0.3",
    "pedalboard>=0.9.22",
    "pyjwt>=2.11.0",
    "pytest>=9.0.2",
    "sqlmodel>=0.0.37",
]
